refactor: Avoid code duplication by moving `getBounds`
The prop3D has three children that were redefining `getBounds` in the same way.
Move the definition of `getBounds` in the prop3D.
Update typescript accordingly.
