+++ Sources/Rendering/Core/ImageSlice/index.d.ts
@@ -18,18 +18,6 @@ export interface vtkImageSlice extends vtkProp3D {
   getActors(): any;
 
   /**
-   * Get the bounds for this mapper as [xmin, xmax, ymin, ymax,zmin, zmax].
-   * @return {Bounds} The bounds for the mapper.
-   */
-  getBounds(): Bounds;
-
-  /**
-   * Get the bounds for this mapper as [xmin, xmax, ymin, ymax,zmin, zmax].
-   * @return {Bounds} The bounds for the mapper.
-   */
-  getBoundsByReference(): Bounds;
-
-  /**
    * Get the bounds for a given slice as [xmin, xmax, ymin, ymax,zmin, zmax].
    * @param {Number} slice The slice index. If undefined, the current slice is considered.
    * @param {Number} [thickness] The slice thickness. If undefined, 0 is considered.
